Crate rune_runtime

Modules

common_capabilities

common_outputs

Structs

Function
A type-erased function which can be used with WebAssembly values.

Signature
The signature for a WebAssembly function.

Enums

FromValuesError

ParameterError

WasmType

WasmValue
A primitive type that can be passed between host and WebAssembly guest.

Traits

CallContext
Contextual information passed to a host function.

Capability

Image

Output
Something a Rune can send output to.

Registrar
A helper type for registering functions and variables that the WebAssembly will be given access to.

WasmTypeList
A list of WebAssembly types, typically used for function arguments or return values.