Trait rune_runtime::CallContext

pub trait CallContext {
    fn memory(&self, address: u32, len: u32) -> Result<&[u8], Error>;

    unsafe fn memory_mut(
        &self, 
        address: u32, 
        len: u32
    ) -> Result<&mut [u8], Error>;

    fn utf8_str(&self, address: u32, len: u32) -> Result<&str, Error> { ... }
}

Contextual information passed to a host function.

Required methods

fn memory(&self, address: u32, len: u32) -> Result<&[u8], Error>

Get immutable access to some WebAssembly memory.

unsafe fn memory_mut(&self, address: u32, len: u32) -> Result<&mut [u8], Error>

Get mutable access to some WebAssembly memory.

Safety

The caller must ensure there are no other references to the returned memory until the slice is dropped.

Loading content...
Provided methods

fn utf8_str(&self, address: u32, len: u32) -> Result<&str, Error>

Read a UTF-8 string from WebAssembly memory.

Loading content...