Trait rune_runtime::Capability

pub trait Capability: Send + Debug + 'static {
    fn generate(&mut self, buffer: &mut [u8]) -> Result<usize, Error>;

    fn set_parameter(
        &mut self, 
        name: &str, 
        value: Value
    ) -> Result<(), ParameterError>;
}

Required methods

fn generate(&mut self, buffer: &mut [u8]) -> Result<usize, Error>

Generate the desired input, writing it to the provided buffer and returning the number of bytes written.

fn set_parameter(
&mut self,
name: &str,
value: Value
) -> Result<(), ParameterError>